New Dev Build

See merge request ultreon/quantum/game!5

+++ client/src/main/resources/assets/quantum/shaders/scene.vert
@@ -1,7 +1,7 @@
-#version 410
+#line 1
 
 #ifdef GL_ES
-precision mediump float;
+precision highp float;
 #endif
 
 in vec3 a_position;
